revise relational model

+++ src/main/java/com/academy/Repositories/ServiceRepository.java
@@ -3,5 +3,5 @@ package com.academy.Repositories;
 import com.academy.Models.Service;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<Service, Integer> {
+public interface ServiceRepository extends JpaRepository<Service, Long> {
 }
