+++ src/main/java/com/academy/Repositories/ServiceTypeRepository.java
@@ -3,5 +3,5 @@ package com.academy.Repositories;
 import com.academy.Models.ServiceType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceTypeRepository extends JpaRepository<ServiceType, Integer> {
+public interface ServiceTypeRepository extends JpaRepository<ServiceType, Long> {
 }
